Add create-prs.sh script
This is a temporary solution to keeping the packages up to date. I'm
hesitant in relying on Renovate for this because, by default, it cannot
distinguish between a package vs dependency update.

If a dependency is updated, but not the actual package, the package will
be rebuilt into the same exact version. If we then release this updated
package, the same version will now have different digests for the wheel,
for example. This would be break consumers.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>