Remove ecosystem and show-sbom tasks
The ecosystem-cert-preflight-checks doesn't seem to be relevant when
building python wheels since the container image built itself is never
released. It's merely a means of carrying the artifacts over to the
release pipeline.

The show-sbom task provides no real value other than spam logs. There
are better ways of getting the sbom.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>