Use hermeto image from prefetch-deps task
This fixes a discrepancy in the `build-locally.sh` script to ensure it
uses the same hermeto image used in the build pipeline.

Also, it auto removes the `podman` containers that it creates, and
ensure the latest image is pulled from the registry which is useful when
using a non-pinned reference.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>